+++ lib/ansible/modules/cloud/amazon/ec2_vol_facts.py
@@ -65,6 +65,8 @@ EXAMPLES = '''
 # fix this
 RETURN = '''# '''
 
+import traceback
+
 try:
     import boto.ec2
     from boto.exception import BotoServerError
@@ -74,6 +76,7 @@ except ImportError:
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.ec2 import connect_to_aws, ec2_argument_spec, get_aws_connection_info
+from ansible.module_utils._text import to_native
 
 
 def get_volume_info(volume):
@@ -136,8 +139,8 @@ def main():
     if region:
         try:
             connection = connect_to_aws(boto.ec2, region, **aws_connect_params)
-        except (boto.exception.NoAuthHandlerFound, StandardError) as e:
-            module.fail_json(msg=str(e))
+        except (boto.exception.NoAuthHandlerFound, Exception) as e:
+            module.fail_json(msg=to_native(e), exception=traceback.format_exc())
     else:
         module.fail_json(msg="region must be specified")
