+++ test/sanity/code-smell/replace-urlopen.sh
@@ -3,7 +3,7 @@
 urllib_users=$(find . -name '*.py' -exec grep -H urlopen '{}' '+' | grep -v \
     -e '^[^:]*/.tox/' \
     -e '^\./lib/ansible/module_utils/urls.py:' \
-    -e '^\./lib/ansible/module_utils/six/_six.py:' \
+    -e '^\./lib/ansible/module_utils/six/__init__.py:' \
     -e '^[^:]*:#'
 )
