+++ test/sanity/pylint/disable.txt
@@ -7,9 +7,9 @@ locally-disabled
 locally-enabled
 method-hidden
 no-member
-no-name-in-module
 no-value-for-parameter
 non-iterator-returned
+no-name-in-module
 not-a-mapping
 not-an-iterable
 not-callable
